+++ src/main/java/pl/xnik3e/Guardian/components/Command/Commands/TogglePrefixCommand.java
@@ -22,7 +22,9 @@ public class TogglePrefixCommand implements ICommand {
 
     @Override
     public void handle(CommandContext ctx) {
-        ctx.getMessage().delete().queue();
+        boolean deleteTriggerMessage = fireStoreService.getModel().isDeleteTriggerMessage();
+        if(deleteTriggerMessage)
+            ctx.getMessage().delete().queue();
         fireStoreService.getModel().setRespondByPrefix(true);
         List<String> args = ctx.getArgs();
         EmbedBuilder eBuilder = new EmbedBuilder();
